refactor: move home components
